가게 상세정보 기능 관련
특정 가게의 모임 리스트 보여주기 UI완성.
선택하면, 해당 모임의 글 보여주기

+++ app/src/main/java/com/musthave0145/mochelins/MapsFragment.java
@@ -180,7 +180,7 @@ public class MapsFragment extends Fragment implements OnMapReadyCallback{
                                         int isviewValue =(int) store.view;
                                         String isViewString = String.valueOf(isviewValue);
 
-                                        int isRatingValue = store.rating;
+                                        double isRatingValue = store.rating;
                                         int isStringInt = (int) isRatingValue;
                                         //가게이름
                                         txtName.setText(store.storeName);
@@ -195,7 +195,7 @@ public class MapsFragment extends Fragment implements OnMapReadyCallback{
                                         //가게 사진
                                         Glide.with(MapsFragment.this).load(store.photo).into(photo);
                                         //별점
-                                        ratingBar.setRating(isRatingValue);
+                                        ratingBar.setRating((float) isRatingValue);
 
                                     } else {
                                         Log.d("ApiResponse", "storeList is null");
